integ-tests: use IAMv3 API key (#13)
We use an IAMv3 key to create the test cluster and the key for the CSI
secret. After each test we clean up the role and key that was created,
if the CSI is torn down.

We also extend the timeout of the tests so that it has enough time to
create the cluster if it does not exist yet.

To make the tests more reliable we wait for Calico and the CSI workloads
to become ready and reduce the amount of requests to avoid rate
limiting.