CI: action to manually destroy integ test cluster (#41)
Sometimes the cluster we use for integration testing can get into a bad
state and integration tests fail. This action allows you to delete the
test cluster from the GitHub UI. As you can see in this PR the
integration tests succeeded again afterwards. (The cluster is recreated
automatically on the next run.