CSI provisioner: Add longer timeout on volumes operations (#99)
# Description


Here is the default timeout on volume operation:

```
  --timeout duration                          Timeout for waiting for volume operation (creation, deletion, capacity queries) (default 10s)
```

This is solving an issue when creating a volume from a large snapshot
(this operation can take very long time)
This prevents provisioner canceling the grpc request context on volume
creation.


<!--
* Prefix: the title with the component name being changed. Add a short
and self describing sentence to ease the review
* Please add a few lines providing context and describing the change
* Please self comment changes whenever applicable to help with the
review process
* Please keep the checklist as part of the PR. Tick what applies to this
change.
-->

## Checklist
(For exoscale contributors)

* [ ] Changelog updated (under *Unreleased* block)
* [ ] Integration tests OK

## Testing

<!--
Describe the tests you did
-->

---------

Signed-off-by: Pierre-Emmanuel Jacquier <15922119+pierre-emmanuelJ@users.noreply.github.com>