feat(node): Add preStop hook (#132)
# Description

When kubernetes drains a node, it will terminate all pods in random
order.
If we get unlucky, and the csi-node agent is terminated before any pods
with VolumeAttachments, we will not be able to detach them. This can
leave pods in state Terminating and can prevent recreation of the pods
because kubernetes considers the underlying volumes still attached.
It requires manual intervention to get out of this state.

Our solution here is to add an preStop hook to the csi-node lifecycle.
If we detect that we are being terminated because the host is being
stopped, we will block until all VolumeAttachments on the current node
are gone.
This ensures the csi-node agent is around to detach all volumes.

Evidently, this only handles graceful shutdowns of nodes. Manual
operator intervention will still be needed in case the underlying node
has an abrupt non-graceful stop.

[sc-114981]

## Checklist
(For exoscale contributors)

* [x] Changelog updated (under *Unreleased* block)
* [x] Integration tests OK

## Testing

* Added an integration test exercising node taint behaviour.