chore(integration): Fail faster
We were using the testify/assert family of functions to assert in the
integration tests. These will mark the test as failed if the assertion
fails, but will not stop the execution of the test.

This is useful if you need to do multiple assertions on a final output.

However when we're asserting system state during test setup, this is
inefficient. If we expect a certain setup fuction to not return an
error, it typically makes no sense to continue executing the test. In
these scenario's we should use the testify/require family of functions,
which will stop test execution on assert failure.